✂️Rename: Mixins 제거, store 분리
