+++ src/helpers/index.js
@@ -1,4 +1,3 @@
 export * from './dom.js';
 export * from './valid.js';
-export * from './store.js';
 export * from './utils.js';
